+++ packages/hardhat-ledger/src/index.ts
@@ -1,6 +1,6 @@
 import { extendProvider } from "hardhat/config";
 import { HardhatNetworkUserConfig } from "hardhat/types";
-import { createLedgerProvider } from "./internal/createLedgerProvider";
+import { createLedgerProvider } from "./internal/create-ledger-provider";
 
 extendProvider(async (provider, config, network) => {
   const networkConfig = config.networks[network] as HardhatNetworkUserConfig;
